Removed unnecessary casting (12)

+++ SF3Lib/Models/Structs/UnknownUInt16Struct.cs
@@ -13,7 +13,7 @@ namespace SF3.Models.Structs {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_valueAddr), displayName: "UInt16 Value", displayFormat: "X4")]
         public ushort Value {
-            get => (ushort) Data.GetUInt16(_valueAddr);
+            get => Data.GetUInt16(_valueAddr);
             set => Data.SetUInt16(_valueAddr, value);
         }
     }
