+++ SF3Lib/Models/Structs/UnknownUInt8Struct.cs
@@ -13,7 +13,7 @@ namespace SF3.Models.Structs {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_valueAddr), displayName: "UInt8 Value", displayFormat: "X2")]
         public byte Value {
-            get => (byte) Data.GetUInt8(_valueAddr);
+            get => Data.GetUInt8(_valueAddr);
             set => Data.SetUInt8(_valueAddr, value);
         }
     }
